third_party/blink/web_tests/external/wpt/css/css-scroll-snap/unreachable-snap-positions-003-expected.txt

This is a testharness.js-based test.
[FAIL] Snaps to the positions defined by the element as much as possible
  assert_equals: expected 10 but got 0
Harness: the test ran to completion.
